fix(mypy): narrow return types via local annotations / bool cast
Addresses no-any-return errors by introducing typed local variables
that force mypy to narrow `json.loads(...)`, `resp.json()[...]`, and
attribute lookups before returning them. Uses `bool(...)` for boolean
returns that came from unparameterized `==` or `.all_passed`.

No behavior changes.

Files touched: eval/{metrics, benchmarks/{human_eval, custom, aimo,
swe_bench}}, core/{token_estimator, tool_result_persister,
compaction_integration, cache_safe_params}, hooks/{validate_path,
auto_test, auto_lint, security_scan, loop_detector}, config/{union,
config_file}, plugins/{dir_loader, manager}, mcp/{client, transport},
lsp/{subprocess, session}, acp/client, permissions/checker,
auth/manager, critic/mixin, env/{remote, docker}, context/mentions,
tools/{browser, parsers/tree_sitter}, server/webhook, rpc/server,
cli/main.

+++ chimera/eval/metrics.py
@@ -26,14 +26,14 @@ def avg_cost(results: list[Any]) -> float:
     """Average cost across eval results."""
     if not results:
         return 0.0
-    return sum(r.cost for r in results) / len(results)
+    return float(sum(r.cost for r in results) / len(results))
 
 
 def avg_steps(results: list[Any]) -> float:
     """Average steps across eval results."""
     if not results:
         return 0.0
-    return sum(r.steps for r in results) / len(results)
+    return float(sum(r.steps for r in results) / len(results))
 
 
 def resolve_rate(results: list[Any]) -> float:
